Put fetchEntityTableData under Network for lack of a better place
