+++ apps/web/pages/space/[id]/entities.tsx
@@ -6,13 +6,10 @@ import { SpaceNavbar } from '~/modules/components/space/space-navbar';
 import { AppConfig } from '~/modules/config';
 import { SYSTEM_IDS } from '~/modules/constants';
 import { Spacer } from '~/modules/design-system/spacer';
-import { Entity } from '~/modules/models/Entity';
 import { Params } from '~/modules/params';
 import { INetwork, Network } from '~/modules/services/network';
 import { StorageClient } from '~/modules/services/storage';
-import { InitialEntityTableStoreParams } from '~/modules/state/entity-table-store';
 import { EntityTableStoreProvider } from '~/modules/state/entity-table-store-provider';
-import { DEFAULT_PAGE_SIZE } from '~/modules/state/triple-store';
 import { Column, Row, Triple } from '~/modules/types';
 
 interface Props {
@@ -90,7 +87,7 @@ export const getServerSideProps: GetServerSideProps<Props> = async context => {
     typeId: initialSelectedType.entityId,
   };
 
-  const { columns, rows } = await fetchEntityTableData({
+  const { columns, rows } = await network.fetchEntityTableData({
     spaceId,
     params,
     config,
@@ -127,102 +124,4 @@ export const fetchSpaceTypeTriples = async (network: INetwork, spaceId: string) 
   });
 
   return triples;
-};
-
-export const fetchEntityTableData = async ({
-  spaceId,
-  params,
-  config,
-}: {
-  spaceId: string;
-  params: InitialEntityTableStoreParams;
-  config: AppConfig;
-}) => {
-  const storage = new StorageClient(config.ipfs);
-  const subgraph = config.subgraph;
-
-  /* To get our columns, fetch the all attributes from that type (e.g. Person -> Attributes -> Age) */
-  const columnsTriples = await new Network(storage, subgraph).fetchTriples({
-    query: '',
-    space: spaceId,
-    first: 100,
-    skip: 0,
-    filter: [
-      { field: 'entity-id', value: params.typeId },
-      { field: 'attribute-id', value: SYSTEM_IDS.ATTRIBUTES },
-    ],
-  });
-
-  /* To get our rows, first we get all of the entity IDs of the selected type */
-  const rowEntityIds = (
-    await new Network(storage, subgraph).fetchTriples({
-      query: params.query,
-      space: spaceId,
-      first: DEFAULT_PAGE_SIZE,
-      skip: params.pageNumber * DEFAULT_PAGE_SIZE,
-      filter: [
-        { field: 'attribute-id', value: SYSTEM_IDS.TYPES },
-        { field: 'linked-to', value: params.typeId },
-      ],
-    })
-  ).triples.map(triple => triple.entityId);
-
-  /* Then we then fetch all triples associated with those entity IDs */
-  const rowTriples = await Promise.all(
-    rowEntityIds.map(entityId =>
-      new Network(storage, subgraph).fetchTriples({
-        space: spaceId,
-        query: '',
-        skip: 0,
-        first: 100,
-        filter: [{ field: 'entity-id', value: entityId }],
-      })
-    )
-  );
-
-  /* Name and Type are the default columns... */
-  const defaultColumns = [
-    {
-      name: 'Name',
-      id: SYSTEM_IDS.NAME,
-    },
-    {
-      name: 'Type',
-      id: SYSTEM_IDS.TYPE,
-    },
-  ];
-
-  /* ...and then we can format our user-defined schemaColumns */
-  const schemaColumns = columnsTriples.triples.map(triple => ({
-    name: Entity.entityName(triple) || triple.value.id,
-    id: triple.value.id,
-  })) as Column[];
-
-  const columns = [...defaultColumns, ...schemaColumns];
-
-  /* Finally, we can build our initialRows */
-  const rows = rowTriples.map(row => {
-    return row.triples.reduce((acc, triple) => {
-      const column = columns.find(column => column.id === triple.attributeId);
-
-      /* If the column doesn't exist, we don't want to add it to the row */
-      if (!column) {
-        return acc;
-      }
-
-      /* Multiple triples are allowed to be displayed in a single column */
-      return {
-        ...acc,
-        [column.id]: {
-          columnId: column.id,
-          triples: [...(acc[column.id]?.triples ?? []), triple],
-        },
-      };
-    }, {} as Row);
-  });
-
-  return {
-    columns,
-    rows,
-  };
 };
